Enhances dashboard with occupancy stats and charts
Improves the dashboard by adding real-time occupancy statistics for different vehicle types, including cars, motorcycles, and trucks.

Introduces interactive charts to visualize parking trends over time and identify peak hours.

Includes monthly subscription information and integrates a Laravel PWA for enhanced user experience.

Fixes the PDF report path and logs price car model retrival.

Refactors the EstacionamentoController to calculate and display current occupancy, monthly member statistics, and historical parking data.

Includes the addition of laravel-websockets

+++ database/migrations/2022_12_15_223130_add_colunm_tipo_car_to_cars_table.php
@@ -13,9 +13,11 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::table('cars', function (Blueprint $table) {
-            $table->string('tipo_car')->default('carro');
-        });
+        if (!Schema::hasColumn('cars', 'tipo_car')) {
+            Schema::table('cars', function (Blueprint $table) {
+                $table->string('tipo_car')->default('carro')->after('preco');
+            });
+        }
     }
 
     /**
@@ -25,8 +27,10 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::table('cars', function (Blueprint $table) {
-            $table->dropColumn('tipo_car');
-        });
+        if (Schema::hasColumn('cars', 'tipo_car')) {
+            Schema::table('cars', function (Blueprint $table) {
+                $table->dropColumn('tipo_car');
+            });
+        }
     }
 };
